# app/docs/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "MyselfDto" : {
        "type" : "object",
        "properties" : {
          "info" : {
            "$ref" : "#/components/schemas/UserInfoDto"
          },
          "locale" : {
            "type" : "string"
          },
          "features" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "capabilities" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        }
      },
      "UserInfoDto" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "domain" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        }
      }
    }
  },
  "paths" : {
    "/internal/users" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/UserInfoDto"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Get Users",
        "tags" : [ "User Resource" ]
      }
    },
    "/internal/users/email/{email}" : {
      "get" : {
        "parameters" : [ {
          "name" : "email",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserInfoDto"
                }
              }
            }
          }
        },
        "summary" : "Get By Email",
        "tags" : [ "User Resource" ]
      }
    },
    "/internal/users/id/{userId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserInfoDto"
                }
              }
            }
          }
        },
        "summary" : "Get By Id",
        "tags" : [ "User Resource" ]
      }
    },
    "/internal/users/myself" : {
      "get" : {
        "parameters" : [ {
          "name" : "ZM_AUTH_TOKEN",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MyselfDto"
                }
              }
            }
          }
        },
        "summary" : "Get Myself",
        "tags" : [ "User Resource" ]
      }
    }
  },
  "info" : {
    "title" : "carbonio-user-management API",
    "version" : "unversioned"
  },
  "servers" : [ {
    "url" : "http://localhost:10000"
  } ]
}